+++ butano/src/bn_affine_bg_attributes_hbe_ptr.cpp.h
@@ -17,16 +17,20 @@ namespace bn
 affine_bg_attributes_hbe_ptr affine_bg_attributes_hbe_ptr::create(
         affine_bg_ptr bg, const span<const affine_bg_attributes>& attributes_ref)
 {
-    int id = hblank_effects_manager::create(attributes_ref.data(), attributes_ref.size(), int(bg.handle()),
-                                            hblank_effects_manager::handler_type::AFFINE_BG_ATTRIBUTES);
+    int id = hblank_effects_manager::create(
+                attributes_ref.data(), attributes_ref.size(), intptr_t(bg.handle()),
+                hblank_effects_manager::handler_type::AFFINE_BG_ATTRIBUTES);
+
     return affine_bg_attributes_hbe_ptr(id, move(bg));
 }
 
 optional<affine_bg_attributes_hbe_ptr> affine_bg_attributes_hbe_ptr::create_optional(
         affine_bg_ptr bg, const span<const affine_bg_attributes>& attributes_ref)
 {
-    int id = hblank_effects_manager::create_optional(attributes_ref.data(), attributes_ref.size(), int(bg.handle()),
-                                                     hblank_effects_manager::handler_type::AFFINE_BG_ATTRIBUTES);
+    int id = hblank_effects_manager::create_optional(
+                attributes_ref.data(), attributes_ref.size(), intptr_t(bg.handle()),
+                hblank_effects_manager::handler_type::AFFINE_BG_ATTRIBUTES);
+
     optional<affine_bg_attributes_hbe_ptr> result;
 
     if(id >= 0)
