+++ butano/src/bn_affine_bg_pd_register_hbe_handler.h
@@ -20,22 +20,22 @@ class affine_bg_pd_register_attributes_hbe_handler
 {
 
 public:
-    static void setup_target(int, iany&)
+    static void setup_target(intptr_t, iany&)
     {
     }
 
-    [[nodiscard]] static bool target_visible(int target_id)
+    [[nodiscard]] static bool target_visible(intptr_t target_id)
     {
         auto handle = reinterpret_cast<void*>(target_id);
         return bgs_manager::hw_id(handle).has_value();
     }
 
-    [[nodiscard]] static bool target_updated(int, iany&)
+    [[nodiscard]] static bool target_updated(intptr_t, iany&)
     {
         return false;
     }
 
-    [[nodiscard]] static uint16_t* output_register(int target_id)
+    [[nodiscard]] static uint16_t* output_register(intptr_t target_id)
     {
         auto handle = reinterpret_cast<void*>(target_id);
         int hw_id = *bgs_manager::hw_id(handle);
@@ -43,7 +43,7 @@ public:
         return reinterpret_cast<uint16_t*>(result);
     }
 
-    static void write_output_values(int, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
+    static void write_output_values(intptr_t, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
     {
         auto attributes_ptr = reinterpret_cast<const affine_bg_mat_attributes*>(input_values_ptr);
 
@@ -53,11 +53,11 @@ public:
         }
     }
 
-    static void show(int)
+    static void show(intptr_t)
     {
     }
 
-    static void cleanup(int)
+    static void cleanup(intptr_t)
     {
         bgs_manager::reload();
     }
@@ -68,22 +68,22 @@ class affine_bg_pd_register_values_hbe_handler
 {
 
 public:
-    static void setup_target(int, iany&)
+    static void setup_target(intptr_t, iany&)
     {
     }
 
-    [[nodiscard]] static bool target_visible(int target_id)
+    [[nodiscard]] static bool target_visible(intptr_t target_id)
     {
         auto handle = reinterpret_cast<void*>(target_id);
         return bgs_manager::hw_id(handle).has_value();
     }
 
-    [[nodiscard]] static bool target_updated(int, iany&)
+    [[nodiscard]] static bool target_updated(intptr_t, iany&)
     {
         return false;
     }
 
-    [[nodiscard]] static uint16_t* output_register(int target_id)
+    [[nodiscard]] static uint16_t* output_register(intptr_t target_id)
     {
         auto handle = reinterpret_cast<void*>(target_id);
         int hw_id = *bgs_manager::hw_id(handle);
@@ -91,18 +91,18 @@ public:
         return reinterpret_cast<uint16_t*>(result);
     }
 
-    static void write_output_values(int, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
+    static void write_output_values(intptr_t, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
     {
         auto int_source = static_cast<const unsigned*>(input_values_ptr);
         auto int_destination = reinterpret_cast<unsigned*>(output_values_ptr);
         memory::copy(*int_source, display::height() / 2, *int_destination);
     }
 
-    static void show(int)
+    static void show(intptr_t)
     {
     }
 
-    static void cleanup(int)
+    static void cleanup(intptr_t)
     {
         bgs_manager::reload();
     }
