+++ butano/src/bn_bg_palettes_transparent_color_hbe_handler.h
@@ -18,16 +18,16 @@ class bg_palettes_transparent_color_hbe_handler
 {
 
 public:
-    static void setup_target(int, iany&)
+    static void setup_target(intptr_t, iany&)
     {
     }
 
-    [[nodiscard]] static bool target_visible(int)
+    [[nodiscard]] static bool target_visible(intptr_t)
     {
         return true;
     }
 
-    [[nodiscard]] static bool target_updated(int, iany&)
+    [[nodiscard]] static bool target_updated(intptr_t, iany&)
     {
         if(optional<palettes_bank::commit_data> commit_data =
                 palettes_manager::bg_palettes_bank().retrieve_commit_data())
@@ -38,22 +38,22 @@ public:
         return false;
     }
 
-    [[nodiscard]] static uint16_t* output_register(int)
+    [[nodiscard]] static uint16_t* output_register(intptr_t)
     {
         return hw::palettes::bg_transparent_color_register();
     }
 
-    static void write_output_values(int, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
+    static void write_output_values(intptr_t, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
     {
         palettes_manager::bg_palettes_bank().fill_hblank_effect_colors(
                     reinterpret_cast<const color*>(input_values_ptr), output_values_ptr);
     }
 
-    static void show(int)
+    static void show(intptr_t)
     {
     }
 
-    static void cleanup(int)
+    static void cleanup(intptr_t)
     {
         palettes_manager::bg_palettes_bank().reload(0);
     }
