+++ butano/src/bn_rect_window_horizontal_boundaries_hbe_handler.h
@@ -18,17 +18,17 @@ class rect_window_horizontal_boundaries_hbe_handler
 {
 
 public:
-    [[nodiscard]] static bool target_visible(int)
+    [[nodiscard]] static bool target_visible(intptr_t)
     {
         return true;
     }
 
-    static void setup_target(int, iany& target_last_value)
+    static void setup_target(intptr_t, iany& target_last_value)
     {
         target_last_value = pair<int, int>();
     }
 
-    [[nodiscard]] static bool target_updated(int target_id, iany& target_last_value)
+    [[nodiscard]] static bool target_updated(intptr_t target_id, iany& target_last_value)
     {
         pair<int, int>& last_value = target_last_value.value<pair<int, int>>();
         pair<int, int> new_value = display_manager::rect_window_hw_horizontal_boundaries(target_id);
@@ -37,12 +37,12 @@ public:
         return updated;
     }
 
-    [[nodiscard]] static uint16_t* output_register(int target_id)
+    [[nodiscard]] static uint16_t* output_register(intptr_t target_id)
     {
         return hw::display::window_horizontal_boundaries_register(target_id);
     }
 
-    static void write_output_values(int, const iany& target_last_value, const void* input_values_ptr,
+    static void write_output_values(intptr_t, const iany& target_last_value, const void* input_values_ptr,
                                     uint16_t* output_values_ptr)
     {
         const pair<int, int>& last_value = target_last_value.value<pair<int, int>>();
@@ -51,11 +51,11 @@ public:
                     last_value, fixed_pairs_ptr, output_values_ptr);
     }
 
-    static void show(int)
+    static void show(intptr_t)
     {
     }
 
-    static void cleanup(int)
+    static void cleanup(intptr_t)
     {
         display_manager::reload_rect_windows_boundaries();
     }
